Organize setup.py with helpers/
